requirements.txt
-i https://pypi.python.org/simple
amqp==2.6.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
attrs==19.3.0
awsebcli==3.18.1
bcrypt==3.1.7
billiard==3.6.3.0
blessed==1.17.8
botocore==1.15.49
cached-property==1.5.1
cairocffi==1.1.0
cairosvg==2.4.2
celery==4.4.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cement==2.8.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
cssselect2==0.3.0
defusedxml==0.6.0
diff-match-patch==20200713; python_version >= '2.7'
django-appconf==1.0.4
django-celery-email==3.0.0
django-ckeditor==5.9.0
django-colorfield==0.3.1
django-countries==6.1.2
django-import-export==2.4.0
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
django==2.2.13
docker-compose==1.25.5
docker[ssh]==4.2.1
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
et-xmlfile==1.0.1
future==0.16.0
html5lib==1.1
idna==2.7
importlib-metadata==1.6.1
jdcal==1.4.1
jmespath==0.10.0
jsonschema==3.2.0
kombu==4.6.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markuppy==1.14
odfpy==1.4.1
openpyxl==2.6.4
paramiko==2.7.1
pathspec==0.5.9
pillow==7.1.2
pipfile==0.0.2
pycparser==2.20
pynacl==1.4.0
pyphen==0.9.5
pyrsistent==0.16.0
python-dateutil==2.8.0
pytz==2020.1
pyyaml==5.3.1
reportlab==3.5.42
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.1
tablib[html,ods,xls,xlsx,yaml]==2.0.0; python_version >= '3.5'
termcolor==1.1.0
texttable==1.6.2
tinycss2==1.0.2
toml==0.10.1
urllib3==1.24.3
vine==1.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
wcwidth==0.1.9
weasyprint==51
webencodings==0.5.1
websocket-client==0.57.0
xlrd==1.2.0
xlwt==1.3.0
zipp==1.2.0; python_version >= '2.7'
